fix: move configuration to separate path
